+++ packages/tx-utils/src/signed-extensions/chain/CheckSpecVersion.ts
@@ -2,5 +2,11 @@ import { MetadataLookup } from "@polkadot-api/metadata-builders"
 import type { SignedExtension } from "../internal-types"
 import { empty, signedExtension, systemVersionProp } from "../utils"
 
-export const CheckSpecVersion = (lookupFn: MetadataLookup): SignedExtension =>
-  signedExtension(empty, systemVersionProp("spec_version", lookupFn))
+export const CheckSpecVersion = (
+  lookupFn: MetadataLookup,
+  specVersion?: number,
+): SignedExtension =>
+  signedExtension(
+    empty,
+    systemVersionProp("spec_version", lookupFn, specVersion),
+  )